refactor: LottoNumberGenerator interface화
