+++ src/test/java/lotto/model/LottosTest.java
@@ -10,7 +10,7 @@ class LottosTest {
     @Test
     @DisplayName("3개의 로또 생성하는지 확인")
     void GenerateLottos() {
-        Lottos lottos = new Lottos(new LottoNumberGenerator(), 3);
+        Lottos lottos = new Lottos(new ShuffleGenerator(), 3);
 
         assertThat(lottos.getLottos().size()).isEqualTo(3);
     }
